refactor formPrefixOption
