+++ src/components/FormPrefixOption.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export default function FormPrefixOption({ id, selected, emoji, prefixText, description }) {
+  return (
+    <option value={id} key={id} selected={selected}>
+      {emoji} {prefixText}: {description}
+    </option>
+  );
+}
