+++ src/data/formModel.tsx
@@ -1,0 +1,11 @@
+const formModel = {
+  type: 'init',
+  prefix: 'init',
+  emoji: 'init',
+  emojiCode: 'init',
+  summary: 'init',
+  description: 'init',
+  issueNumber: '',
+};
+
+export default formModel;
